% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_row_df.R
\name{add_row_df}
\alias{add_row_df}
\title{add_row_df}
\usage{
add_row_df(x, add_row = NULL, pos = NULL)
}
\arguments{
\item{x}{data.frame}

\item{add_row}{named vector}

\item{pos}{an rbind_at}
}
\value{
data.frame
}
\description{
Add row to data.frame.
}
\examples{
df <-   data.frame(
Source = c("A", "B", "C", "F"),
x = 1:4,
y = 1:4,
stringsAsFactors = FALSE
)

add_row_df(df, c("Erste Zeile" = 1, "Dritte" = 3))

add_row_df(df, "Erste Zeile")
add_row_df(df, c("Erste Zeile", "Zweite"))
add_row_df(df, c("Erste Zeile" = 1, "letzte" = 5))



}
