% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to.R
\name{add_value}
\alias{add_value}
\title{Manually change values in data.frames}
\usage{
add_value(x, row, col, value = NA)
}
\arguments{
\item{x}{data.fram to be canged.}

\item{row}{which row or which condition sex == "male"}

\item{col}{which column}

\item{value}{the value to be changed}
}
\value{
data.frame
}
\description{
Manually change values in data.frames
}
\examples{

DF <- data.frame(
pat.id = 1:5,
sex = factor(c("male", NA, "female", "male", "female")),
hight = c(170, 178, 185, 169, 10)
)

add_value(DF, 5, 3, NA)
add_value(DF, pat.id == 2, sex, "female")
add_value(DF, hight < 100, hight, 175)

}
