% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate-auc_trapezoid.R
\name{auc_trapezoid}
\alias{auc_trapezoid}
\title{Area Under the Curve}
\usage{
auc_trapezoid(..., time = NULL, na.rm = FALSE)
}
\arguments{
\item{...}{value}

\item{time}{time}

\item{na.rm}{na.rm}
}
\value{
vector
}
\description{
Method trapezoid
}
\examples{

 dat <-
data.frame(
  t0 = c(1, 1, 1),
  t2 = c(1, 2, 3),
  t3 = c(1, 3, 5),
  t4 = c(1, 3, 6)
)

#apply(dat, 1, function(x)
 # sum(diff(time) * zoo::rollmean(x, 2)))
time = c(0, 1, 2, 3)
auc_trapezoid(1,1,1,1)
auc_trapezoid(dat )
transform(dat, auc= auc_trapezoid( t0, t2, t3, t4, time= time))

}
