% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-auto-trans.R
\name{auto_trans}
\alias{auto_trans}
\title{transformation}
\usage{
auto_trans(x, treshhold = 1)
}
\arguments{
\item{x}{Vector}

\item{treshhold}{grundlienie}
}
\value{
Vector mit attr link, inverse und name
}
\description{
transformation
}
\examples{
n<-100
x1 = rnorm(100)+5
x2 = rlnorm(n, meanlog = 0, sdlog = 1)+1
x3 = rpois(n, lambda = 1)+1
x4 = rweibull(n, shape = .8, scale = 1)+10
x5 = runif(n, min = 0, max = 100)+1


auto_trans(x2)
#'
x.neg <- rbeta(10000, 5, 2)  # Negative Skew

x.pos <- rexp(1000, 1) # Positive Skew
par(mfrow=c(2,2))
hist(x.neg)
hist(x.pos)
hist(auto_trans(x.neg))
hist(auto_trans(x.pos))

}
