% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string-clean-names.R
\name{clean_names}
\alias{clean_names}
\alias{clean_names2}
\alias{clean_names.data.frame}
\alias{clean_names.default}
\alias{paste_names}
\alias{paste_names_levels}
\alias{cleansing_umlaute}
\alias{cleansing_umlaute2}
\alias{abbreviate2}
\title{clean_names}
\usage{
clean_names(x, ...)

clean_names2(
  x,
  tolower = TRUE,
  abbreviate = TRUE,
  minlength = 25,
  replace = c(`\%` = "_pct", `#` = "_cnt", `¨` = "", `&+` = "_and_", `@+` = "_at_", `_`
    = ".")
)

\method{clean_names}{data.frame}(
  x,
  label = TRUE,
  labels = NULL,
  cleanup.encoding = FALSE,
  from = "UTF8",
  to = "latin1",
  ...
)

\method{clean_names}{default}(
  x,
  tolower = TRUE,
  unique = TRUE,
  abbreviate = FALSE,
  minlength = 4,
  replace = c(`'` = "", `"` = "", `\%` = "_pct", `#` = "_cnt", ä = "ae", ü = "ue", ö =
    "oe", Ü = "Ue", Ä = "Ae", Ö = "Oe", ß = "ss", `¨` = "", `&+` = "_and_", `@+` =
    "_at_", `_` = "."),
  ...
)

paste_names(x, collapse = ", ", ...)

paste_names_levels(x, abbreviate = TRUE, collapse = ", ", ...)

cleansing_umlaute(x, ...)

cleansing_umlaute2(x, ...)

abbreviate2(x, minlength = 4)
}
\arguments{
\item{x}{string}

\item{...}{alles}

\item{tolower}{alles zu kleinbuschstaben}

\item{abbreviate, minlength}{Abbkürzung}

\item{replace}{named character string}

\item{label, labels}{label}

\item{cleanup.encoding, from, to}{logical  UTF-8 to latin}

\item{unique}{eindeitige namen}

\item{collapse}{an  character string to separate the results ", "}

\item{data}{data.frame oder character}
}
\value{
die selbe Klasse wie der Input

character

character
}
\description{
Alternative zu `janitor`
Quelle https://drdoane.com/clean-consistent-column-names/

clean_names2() ist eine einfachere Variante von clean_names(). 
Die Funktion gibt es nur für Vektoren. Die Umlaute werden nicht mit Ue oder oe
übersätzt.

clean_names(): Input data.frame  output ist ein 
data.frame mit bereinigten namen.

paste_names:  paste names

paste_names_levels:  paste names + Labels Die funktion ist gedacht für 
Copy and Paste.

cleansing_umlaute(): 
Funktion entfernt stoerende Umlaute, unten stehende Liste ggf. erweitern.

cleansing_umlaute2: Sonderzeichen aus socisurvy
}
\examples{


clean_names(tibble::tibble("Öli"=1:3, "p-k"=1:3, "95\%-CI"=4:6) )

c(
  c(
    "  a", "a  ", "a \%", "a",
    "$a", "$$$a", "GDP ($)",
    "GDP (us$)", "a (#)", "a & b",
    "#", "$", "a_cnt", "Aa&Bb",
    "camelCasePhrases", "AlphaBetaGamma",
    "Alpha       Beta", "Beta  !!! Gamma",
    "a + b", "a - b", "a * b", "Ösel"
  ), abbreviate=TRUE
)



#' 
df <- data.frame(
  Öse = c(1, 2, 3, 1, 2, 3),
  Löre = gl(2, 3, labels = c("Amy", "Bob")),
  Fürn = c(9, 7, 6, 8, 6, 9),
  Mäße = c(6, 7, 8, 5, 6, 7),
  hüne=c(1, 2, 3, 1, 2, 3)
)

# Beide funktionen machen das gleiche 
# nur fix_names ist etwas einfacher gestrickt.
get_label(clean_names(df))

get_label(fix_names(df))

abbreviate2(c("Hallö", "Halü"))

}
