% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data2.R
\name{cleaup_formula}
\alias{cleaup_formula}
\title{Texte also Überschfifte werden zu logical mit NA
daher hie die Heder vergeben}
\usage{
cleaup_formula(formula, data, groups)
}
\description{
Texte also Überschfifte werden zu logical mit NA
daher hie die Heder vergeben
}
