% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cs.R
\name{Cs}
\alias{Cs}
\alias{XLS}
\title{Character strings from unquoted names}
\usage{
Cs(...)

XLS(...)
}
\arguments{
\item{...}{liste mit den Spaltennamen A:BB}
}
\value{
Vektor
}
\description{
Erweiterung der  Hmisc Cs-Funktion um trim.ws. Erlaubt Variablen mit Komma oder Plus(+) oder Leerzeichen abzuternnen.

XLS:

Nummernindex aus Excel - Spaltenbezeichnung

Extrahiert aus Buchstaben die Spaltennummer
}
\examples{

Cs(sd, fr, fg)
Cs(sd, fr, fg, "hju nh")
Cs("  Hallo Welt ")
Cs(~ sd + fr + fg)
as.roman(1968)
#strsplit("A:V", "\\:")
XLS(a, B)
XLS(a, B, c:f, g:h,i, r:z)
XLS(A:Z)

}
