% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor2.R
\name{factor2}
\alias{factor2}
\alias{reorder2}
\title{Factors}
\usage{
factor2(
  x,
  ...,
  levels,
  labels,
  exclude = NA,
  ordered = is.ordered(x),
  nmax = NA
)

reorder2(x, X, ..., decreasing = TRUE, last = NULL)
}
\arguments{
\item{x}{a vector of data,}

\item{levels, labels, exclude, ordered, nmax}{an die Funktion factor}

\item{X, ...}{an reorder}

\item{decreasing}{an ordere logical}

\item{last}{character}
}
\value{
factor
}
\description{
Fancy copie der fun factor.
}
\examples{


x <- c(1, 0, 0, 0, 1, 1, 0, 3, 2, 2)

x <- data.frame(x = x,
                sex = factor2(
                  x,
                  male = 1,
                  female = 0,
                  div = 3,
                  other = 2
                ))
x
levels(x$sex)


table(reorder2(x$sex, last = "other"))


#' lattice::barchart(rev(reorder2(x$sex, last = "other")))

# dat<-as.data.frame(table(x$sex)) 
# lattice::barchart( 
#   reorder2(Var1, Freq, last="other") ~Freq, 
#   dat, 
#   origin =0)
# 
# with(
#   dat,
#   reorder2(Var1, Freq, last="other"))
#'
}
