% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_to_dummy.R
\name{factor_to_dummy}
\alias{factor_to_dummy}
\alias{dummy_to_factor}
\title{Creation of dummy variables and reverse}
\usage{
factor_to_dummy(x)

dummy_to_factor(x, value = 1L)
}
\arguments{
\item{x}{data.frame all 0 / 1}
}
\value{
data.frame

factor
}
\description{
Creation of dummy variables and reverse
}
\examples{

 z <- gl(3, 2, 12, labels = c("apple", "salad", "orange"))
table(z)
levels(z) <- list("veg"   = "salad", "fruit" = c("apple", "orange"))
table(z)
z <- factor_to_dummy(z)
table(z)
z <- dummy_to_factor(z)
table(z)
}
