% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter2}
\alias{filter2}
\title{Keep rows that match a condition}
\usage{
filter2(.data, ..., .preserve = FALSE, subject_column_name = NULL)
}
\arguments{
\item{.data}{A data frame,}

\item{...}{Expressions}

\item{.preserve}{Relevant when the .data input is grouped.}

\item{subject_column_name}{Unique subject id}
}
\value{
data.frame mit attr(x, "filter")
}
\description{
copy from filter {dplyr} and https://github.com/openpharma/visR
}
\examples{

 #?prepare_consort
#'
require(stp25stat2)
require(stp25tools)


data(DFdummy, package = "stp25data")

DF1 <- DFdummy \%>\% filter2(study.agreement)
attr(DF1, "filter")

DF2 <- DF1 \%>\% filter2(
  st.p.sars.cov2 == "nein",
  !is.na(spike.igg.3.impfung),
  !is.na(MPN)

)

DF3 <- DF2 \%>\% filter2(
  study.agreement,
  sero.negativ.after.dose.2,
  !is.na(spike.igg.3.impfung),
  !is.na(spike.igg.4.impfung),
  spike.igg.3.impfung == "<7.1 BAU/ml"
)



dat <- prepare_consort(DF1, DF2, DF3)


require(consort)

out <- consort_plot(
  data = dat,
  orders = c(
    Trial.Nr   = "Population",
    Condition.1           = "Excluded",
    Trial.Nr     = "Allocated \nDeskriptive Analyse",
    Condition.2    =    "Fehlende Daten",
    Trial.Nr = "Regressionsanalyse",
    Condition.3    = "Not evaluable for the final analysis",
    Trial.Nr = "Final Analysis"
  ),
  side_box = c("Condition.1", "Condition.2", "Condition.3"),
  cex = 0.9
)



plot(out)
}
