% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation-lod.R
\name{imputation_LOD}
\alias{imputation_LOD}
\alias{imputation_LOD.default}
\alias{imputation_LOD.data.frame}
\title{Imputation of LOD}
\usage{
imputation_LOD(x, ...)

\method{imputation_LOD}{default}(
  x,
  lod = NULL,
  threshold = 0.0707,
  na.replace = TRUE,
  force.lod = TRUE,
  accuracy = 1
)

\method{imputation_LOD}{data.frame}(
  x,
  lod = NULL,
  threshold = 0.0707,
  na.replace = TRUE,
  force.lod = TRUE,
  accuracy = 1,
  remove.lod = 1/2
)
}
\arguments{
\item{x}{numeric  vector to round}

\item{lod}{Limit of detection if NULL lod = min - 3sd}

\item{threshold}{default 1/sqrt(2) replace NA and x < lod by lod * threshold}

\item{na.replace}{should NA be replaced by the lod  * threshold}

\item{force.lod}{should all values smaller than the LOD be replaced?}

\item{accuracy}{significant digits for rounding}

\item{remove.lod}{NULL or fraction}
}
\value{
vector, data.frame
}
\description{
Limit Of Detection (LOD)  is the minimum detectable value. (blank signal + 3SD)
}
\details{
Replace NA by LOD*threshold and
round to the next order of magnitude

https://github.com/jranke/chemCal
}
\examples{

 x <-
c(.00049001,.0035648,.01,.0112,
  .023212548,.00541257,.004041257,.458,.500)
y <-
  c(43.01,49.156,678.00112458964,789.023212548,
    674.00049001,634.00541257,76.004041257,789.458,500
  )

data.frame(x = signif(x, 3), x.lod = imputation_LOD(x,  lod = .0035648))

data.frame(y = signif(y, 3), y.lod = imputation_LOD(y,  lod = 49.156))


#' #require(stp25tools)

DF <-
  data.frame(x=rnorm(10), y=rnorm(10), z=rnorm(10)) |>
  Label( x ="Asp", y ="Trp", z ="Leu")
DF[1,1] <- NA

imputation_LOD(DF, lod =c(-.2, .2, .05), accuracy=3)
imputation_LOD(DF, lod = c(-.2, .2, .05), accuracy = 3, na.replace =FALSE)
}
