% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label-data-attributes.R
\name{Label}
\alias{Label}
\alias{delet_label}
\alias{set_label}
\alias{trimm_label}
\alias{gsub_label}
\alias{get_label}
\alias{label_data_frame}
\title{Label}
\usage{
Label(data, ...)

delet_label(data)

set_label(data, label = NULL, include.units = FALSE)

trimm_label(data, pattern = "\\\\.\\\\.\\\\. ", pos = 2)

gsub_label(data, pattern = "\\\\&amp;", replacement = "&")

get_label(data, ..., include.units = FALSE)

label_data_frame(data, label)
}
\arguments{
\item{data}{data.frame}

\item{...}{label in der Form a="Hallo, b="Welt"}

\item{label}{label als named vector}

\item{include.units}{einheiten}

\item{pattern}{Pattern to look for.}
}
\description{
Setzen des attr(data, "label")

delet_label:  Loeschen aller Attributs label

set_label:  Setzen  der Attributs label

get_label und GetLabelOrName:  Abrufen  der Attributs label

Intern wenn mit get_label nur die Kopie wiederhergestellt wird.
}
\examples{

df <- data.frame(
BMI=c(1,2,3,1,2,3),
WHtR= gl(2,3, label =c("Amy", "Bob")),
WHtR_1=c(9,7,6,8,6,9),
bildprof=c(6,7,8,5,6,7)
)

DF<-
  Label(df, BMI = "Body-Mass-Index",
        WHtR =  "Waist-Height-Ratio",
        WHtR_1 ="Waist-Height-Ratio"
  )

# DF$BMI<- units::set_units(DF$BMI, kg/m2)

get_label(DF)
get_label(DF, include.units=TRUE)
DF<- set_label(DF, c(bildprof = "Bildungsprofil"))
get_label(DF)
DF<- delet_label(DF)
get_label(DF)

DF<- wrap_label(DF)
get_label(DF)

}
