% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-na-approx.R
\name{na_approx}
\alias{na_approx}
\alias{na_approx.default}
\alias{na_approx.data.frame}
\alias{na_approx.matrix}
\title{Replace NA by Interpolation}
\usage{
na_approx(x, ...)

\method{na_approx}{default}(
  x,
  fun = function(z) {
     zoo::na.approx(z, na.rm = FALSE)
 },
  min.length = 3L
)

\method{na_approx}{data.frame}(x, ...)

\method{na_approx}{matrix}(x)
}
\arguments{
\item{x}{an object.}

\item{...}{further arguments passed to methods.}

\item{fun}{Interpolation  zoo::na.approx(z)  zoo::na.fill(z, "extend")  or}

\item{min.length}{min length}
}
\value{
data.frame
}
\description{
Replace NA by Interpolation
}
\examples{

dat <- data.frame(
  u = c(NA, 2,  3,  4,  6,  NA, 7, NA),
  v = c(NA, NA, NA, NA, 6,  4,  7, 2),
  w = c(NA, 2,  NA, 4,  6,  4,  7, 2),
  x = c(NA, 2,  3,  7,  6,  4,  7, NA),
  y = c(NA, 7,  2,  NA, 1,  9,  3, 4),
  z = c( 1, 7,  2,  8,  NA, 9,  3, 4)
)

na_approx(dat$z)
na_approx(dat)
na_approx(as.matrix(dat))

}
