% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare-consort-plot.R
\name{prepare_consort}
\alias{prepare_consort}
\title{prepare_consort}
\usage{
prepare_consort(..., split = NULL)
}
\arguments{
\item{...}{data.frame mit filter2}

\item{split}{table-Objekt}
}
\value{
data.frame
}
\description{
prepare_consort
}
\examples{
\dontrun{
require(stp25stat2)
require(stp25tools)


data(DFdummy, package = "stp25data")
if(exists("DFdummy")){
  DF1 <- DFdummy |> filter2(study.agreement)
  attr(DF1, "filter")
  
  DF2 <- DF1 |> filter2(
    st.p.sars.cov2 == "nein",
    !is.na(spike.igg.3.impfung),
    !is.na(MPN)
    
  )
  
  DF3 <- DF2 |> filter2(
    study.agreement,
    sero.negativ.after.dose.2,
    !is.na(spike.igg.3.impfung),
    !is.na(spike.igg.4.impfung),
    spike.igg.3.impfung == "<7.1 BAU/ml"
  )
  
  
  
  dat <- prepare_consort(DF1, DF2, DF3)
  
  
  require(consort)
  
  out <- consort_plot(
    data = dat,
    orders = c(
      Trial.Nr   = "Population",
      Condition.1           = "Excluded",
      Trial.Nr     = "Allocated \nDeskriptive Analyse",
      Condition.2    =    "Fehlende Daten",
      Trial.Nr = "Regressionsanalyse",
      Condition.3    = "Not evaluable for the final analysis",
      Trial.Nr = "Final Analysis"
    ),
    side_box = c("Condition.1", "Condition.2", "Condition.3"),
    cex = 0.9
  )
  
  
  
  plot(out)
}


# DF1 <- DFdummy |> filter2(study.agreement)
# attr(DF1, "filter")
# DF2 <- DF1 |> filter2(
#   st.p.sars.cov2 == "nein",
#   !is.na(spike.igg.3.impfung),
#   !is.na(MPN)
# 
# )
# 
# DF3 <- DF2 |> filter2(
#   study.agreement,
#   sero.negativ.after.dose.2,
#   !is.na(spike.igg.3.impfung),
#   !is.na(spike.igg.4.impfung),
#   spike.igg.3.impfung == "<7.1 BAU/ml"
# )
# dat <- prepare_consort(DF1, DF2, DF3)
# 
# require(consort_plot)
# 
# out <- consort_plot(
#   data = dat,
#   orders = c(
#     Trial.Nr   = "Population",
#     Condition.1           = "Excluded",
#     Trial.Nr     = "Allocated \nDeskriptive Analyse",
#     Condition.2    =    "Fehlende Daten",
#     Trial.Nr = "Regressionsanalyse",
#     Condition.3    = "Not evaluable for the final analysis",
#     Trial.Nr = "Final Analysis"
#   ),
#   side_box = c("Condition.1", "Condition.2", "Condition.3"),
#   cex = 0.9
# )
# 
# plot(out)
# 
# 
# dat2 <- prepare_consort(DF1,
#                         DF2,
#                         split = table(DF2$geschlecht)
# )
# 
# consort_plot(
#   data = dat2,
#   orders = c(
#     Trial.Nr   = "Population",
#     Condition.1           = "Excluded",
#     Trial.Nr     = "Allocated \nDeskriptive Analyse",
#     Condition.2    =    "Fehlende Daten",
#     # Trial.Nr = "Regressionsanalyse",
#     Split = "Stratifizierung",
#     Trial.Nr = "Final Analysis"
#   ),
#   side_box = c("Condition.1", "Condition.2"),
# 
#   allocation = "Split",
#   labels = c("1" = "Screening", "2" = "Randomization", "4" = "Final"),
#   cex = 0.9
# )

}


}
