% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{reshape2_colsplit}
\alias{reshape2_colsplit}
\title{Split a vector into multiple columns}
\usage{
reshape2_colsplit(string, pattern, names)
}
\arguments{
\item{string}{character vector or factor to split up}

\item{pattern}{regular expression to split on}

\item{names}{names for output columns}
}
\description{
Stolen from reshape2
}
\examples{
x <- c("a_1", "a_2", "b_2", "c_3")
vars <- reshape2_colsplit(x, "_", c("trt", "time"))
vars
str(vars)
}
