% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lod.R
\name{round_lod}
\alias{round_lod}
\alias{round_lod.default}
\alias{round_lod.data.frame}
\title{Round to the next order of magnitude}
\usage{
round_lod(x, ...)

\method{round_lod}{default}(x, lod = NULL, threshold = 0.5, na.replace = TRUE, accuracy = 1)

\method{round_lod}{data.frame}(x, lod = NULL, threshold = 0.5, na.replace = TRUE, accuracy = 1)
}
\arguments{
\item{x}{numeric  vector to round}

\item{lod}{Limit of dedection (accuracy)}

\item{threshold}{replace NA and x < lod by lod * threshold}

\item{na.replace}{should NA be replaced by the lod  * threshold}

\item{accuracy}{significant digits for rounding}
}
\value{
vector, data.frame
}
\description{
Round to the next order of magnitude
}
\examples{

 x <-
c(.00049001,.0035648,.01,.0112,
  .023212548,.00541257,.004041257,.458,.500)
y <-
  c(43.01,49.156,678.00112458964,789.023212548,
    674.00049001,634.00541257,76.004041257,789.458,500
  )


data.frame(x = signif(x, 3), x.lod = round_lod(x,  lod = .0035648))

data.frame(y = signif(y, 3), y.lod = round_lod(y,  lod = 49.156))


#' #require(stp25tools)

DF <-
  data.frame(x=rnorm(10), y=rnorm(10), z=rnorm(10)) |>
  Label( x ="Asp", y ="Trp", z ="Leu")
DF[1,1] <- NA

round_lod(DF, lod =c(-.2, .2, .05), accuracy=3)
round_lod(DF, lod = c(-.2, .2, .05), accuracy = 3, na.replace =FALSE)
}
