% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_multiple_choice.R
\name{separate_multiple_choice}
\alias{separate_multiple_choice}
\title{separate multiple choice}
\usage{
separate_multiple_choice(
  x,
  sep = ", ",
  id = 1:length(x),
  out_levels = 1:0,
  out_labels = NULL,
  as_logical = is.null(out_labels),
  prafix = NULL,
  into = NULL,
  label = NULL,
  na.strings = NULL
)
}
\arguments{
\item{x}{Vektor ist entweder ein Character oder eine zahl
Bei Zahlen gehen nur 1 bis 11}

\item{sep}{wichtig ist das trennzeichen}

\item{out_labels, out_levels}{Outut nimmeric oder factor}

\item{as_logical}{ausgabe logical}

\item{prafix}{neue Variablen Name fuer die ausgabe default ist der Name vom Input}

\item{into}{brauch nicht gesetzt werden entspricht der Anzahl an levels}

\item{label}{Labels}
}
\value{
data.frame mit 0 1 Kodierung und den Fragen als Labels
}
\description{
Aufdroeseln von Mehrfachantworten
}
\details{
\code{separate_multiple_choice()} Aufdröseln vom Mehrfachantworten, die Funktion transformiert einen String mit Trennzeichen
zu einem Multi-Set mit 0 und 1. (Separate multiple choice)
}
\examples{

 #require(stp25tools)
lbl <- c (
  '1 Heart Failure' = 1,
  '2 Rhythm Abnormality' = 2,
  '3 Valve Dysfunction' = 3,
  '4 Bleeding with OAC' = 4,
  '5 ACS' = 5,
  '6 Neurological Event' = 6,
  '7 Neoplastic Disease' = 7,
  '8 Others' = 8,
  '0 No Complications' = 0
)

x <- c(0,
       "1,3,6,8,4",
       2,
       "",
       "8,4",
       #  3.8,100,
       "4,6,8,3",
       "2,3,4,5")

#x <- gsub("\\.", ",", x)
rslt <-
  separate_multiple_choice(x ,
                           sep = ",",
                           as_logical = TRUE,
                           label = lbl)
stp25stat2::Tbll_desc(rslt)



dat <-  data.frame(Q1=c(134, NA, 35, 134, 5, 24),
                   Q2=c(
                     "Alm Dudler, Essig, Cola",
                     NA,
                     "Cola, Holer",
                     "Alm Dudler, Cola, Essig",
                     "Holer",
                     "Bier, Essig"))



dat2 <- cbind(dat[2],
              separate_multiple_choice(dat$Q2))
names(dat2) <- get_label(dat2)
dat2


dat1 <- cbind(dat[1],
              separate_multiple_choice(
                dat$Q1,
                label = 
                  c("Alm Dudler", 
                    "Bier", 
                    "Cola", 
                    "Essig", 
                    "Holer")))
names(dat1) <- get_label(dat1)
dat1


x<-data.frame(A=c(15911, 261011, 3711, 48111))
separate_multiple_choice(x$A)
}
