% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-filter.R
\name{simple_consort_plot}
\alias{simple_consort_plot}
\title{consort_plot}
\usage{
simple_consort_plot(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{an  stp25plot:::visr}
}
\value{
plot
}
\description{
consort_plot
}
