% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-subset.R
\name{subset2}
\alias{subset2}
\title{Subsetting Vectors, Matrices and Data Frames}
\usage{
subset2(x, ...)
}
\arguments{
\item{x}{object to be subsetted.}

\item{...}{further arguments to be passed to or from other methods.}
}
\value{
data.frame
}
\description{
Subsetting Vectors, Matrices and Data Frames
}
\examples{

subset2(airquality, Temp > 80, select = c(Ozone, Temp))

}
