% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{visR_get_attrition}
\alias{visR_get_attrition}
\title{Generate cohort attrition table}
\usage{
visR_get_attrition(
  data,
  criteria_descriptions = paste0(seq_along(criteria_conditions), ". Filter"),
  criteria_conditions,
  subject_column_name
)
}
\arguments{
\item{data}{Dataframe. It is used as the input data to count the subjects
that meets the criteria of interest}

\item{criteria_descriptions}{\code{character} It contains the descriptions
of the inclusion/exclusion criteria.
Each element of the vector corresponds to the description of each criterion.}

\item{criteria_conditions}{\code{character} It contains the corresponding
conditions of the criteria.
These conditions will be used in the table to compute the counts of the
subjects.}

\item{subject_column_name}{\code{character} The column name of the table that
contains the subject id.}
}
\value{
The counts and percentages of the remaining and excluded subjects
for each step of the cohort selection in a table format.
}
\description{
Gestolen von  https://github.com/openpharma/visR
}
\details{
Geändert sind die Variablen-Namen
}
