% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor2.R
\name{factor2}
\alias{factor2}
\alias{reorder2}
\title{Factors}
\usage{
factor2(
  x,
  ...,
  levels,
  labels,
  exclude = NA,
  ordered = is.ordered(x),
  nmax = NA
)

reorder2(
  x,
  X,
  ...,
  decreasing = TRUE,
  last = NULL,
  threshold = NULL,
  threshold.na.strings = "Other"
)
}
\arguments{
\item{x}{a vector of data,}

\item{levels, labels, exclude, ordered, nmax}{an die Funktion factor}

\item{X, ...}{an reorder  x => an atomic vector, usually a factor
X    => a vector of the same length as x, whose subset of values for each unique level of x determines the eventual order of that level.}

\item{decreasing}{an ordere logical}

\item{last}{character}

\item{threshold, threshold.na.strings}{Anzahl an minimalen Nennungen}
}
\value{
factor
}
\description{
Fancy copie der fun factor.
}
\examples{


x <- c(1, 0, 0, 0, 1, 1, 0, 3, 2, 2)

x <- data.frame(x = x,
                sex = factor2(
                  x,
                  male = 1,
                  female = 0,
                  div = 3,
                  other = 2
                ))
x
levels(x$sex)


table(reorder2(x$sex, last = "other"))


#' lattice::barchart(rev(reorder2(x$sex, last = "other")))

# dat<-as.data.frame(table(x$sex))
# lattice::barchart(
#   reorder2(Var1, Freq, last="other") ~Freq,
#   dat,
#   origin =0)
#
# with(
#   dat,
#   reorder2(Var1, Freq, last="other"))
#'

 x <-
c(
  rep(1, 21),rep(2, 120),rep(3, 28),rep(4, 4),rep(5, 56),
  rep(6, 2),rep(7, 92),rep(8, 42),rep(9, 74),rep(10, 20)
)

x <- factor(x, 1:10, letters[1:10])
table(x)
table(reorder2(x))
table(reorder2(x, threshold = 30))

}
