% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter2}
\alias{filter2}
\title{Keep rows that match a condition}
\usage{
filter2(.data, ..., .preserve = FALSE, subject_column_name = NULL)
}
\arguments{
\item{.data}{A data frame,}

\item{...}{Expressions}

\item{.preserve}{Relevant when the .data input is grouped.}

\item{subject_column_name}{Unique subject id}
}
\value{
data.frame mit attr(x, "filter")
}
\description{
copy from filter {dplyr} and https://github.com/openpharma/visR
}
