% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Index.R
\name{Index}
\alias{Index}
\alias{Index.data.frame}
\alias{Index.default}
\alias{Sum2}
\title{Summen-Index
 
 Summen Index eine Summenfunktion mit der Erweiterung zum Umcodieren}
\usage{
Index(x, ...)

\method{Index}{data.frame}(
  x,
  ...,
  re.code = FALSE,
  fun = "mean",
  na.rm = TRUE,
  digits = 4,
  max.level = NA,
  min.level = NA
)

\method{Index}{default}(
  ...,
  re.code = FALSE,
  fun = "mean",
  na.rm = TRUE,
  digits = 4,
  max.level = NA,
  min.level = NA
)

Sum2(
  ...,
  re.code = FALSE,
  fun = "mean",
  na.rm = TRUE,
  digits = 4,
  max.level = NA,
  min.level = NA
)
}
\value{
Vektor
}
\description{
Summen-Index
 
 Summen Index eine Summenfunktion mit der Erweiterung zum Umcodieren
}
