% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Index.R
\name{Index}
\alias{Index}
\alias{Index.formula}
\alias{Index.data.frame}
\alias{Index.default}
\alias{Sum2}
\title{Summen-Index
 
Summen Index eine Summenfunktion mit der Erweiterung zum Umcodieren}
\usage{
Index(x, ...)

\method{Index}{formula}(x, data, ...)

\method{Index}{data.frame}(
  x,
  ...,
  re.code = FALSE,
  fun = "mean",
  na.rm = TRUE,
  digits = 4,
  max.level = NA,
  min.level = NA
)

\method{Index}{default}(
  ...,
  re.code = FALSE,
  fun = "mean",
  na.rm = TRUE,
  digits = 4,
  max.level = NA,
  min.level = NA
)

Sum2(
  ...,
  re.code = FALSE,
  fun = "mean",
  na.rm = TRUE,
  digits = 4,
  max.level = NA,
  min.level = NA
)
}
\arguments{
\item{x, data}{data.frame x ist entwerde formula oder ein data.frame}

\item{...}{alles weitere}

\item{re.code}{revers code logical}

\item{fun}{mean or sum}

\item{na.rm}{NA remove}

\item{digits}{Nachkomastellen}

\item{max.level, min.level}{nin und max levels}
}
\value{
Vektor
}
\description{
Summen-Index
 
Summen Index eine Summenfunktion mit der Erweiterung zum Umcodieren
}
