% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot-long.R, R/pivot-wide.R
\name{Long}
\alias{Long}
\alias{Long.formula}
\alias{Long.data.frame}
\alias{Long_rbind}
\alias{Wide}
\title{Long und Wide}
\usage{
Long(x, ...)

\method{Long}{formula}(x, data, key = "variable", value = "value", use.label = TRUE, ...)

\method{Long}{data.frame}(
  x,
  ...,
  by = NULL,
  key = "variable",
  value = "value",
  id.vars = all.vars(by),
  use.label = TRUE,
  .list = NULL
)

Long_rbind(
  .data,
  .list = list(NULL),
  by = NULL,
  .id = "time",
  names = NULL,
  .first.data = NULL,
  ...
)

Wide(
  data,
  key,
  ...,
  values_fill = NULL,
  names_sep = "_",
  names_vary = "fastest"
)
}
\arguments{
\item{x}{data.frame oder formula}

\item{...}{key +  values_fro or erster Teil der Formula}

\item{data}{dataframe}

\item{key, value}{Namen fuer die Ausgabe}

\item{use.label}{attribut label verwenden}

\item{by, id.vars}{Items}

\item{.list}{Spalten die aufgedröselt werden list(t0 = 1:3, t1 = 4:6, t2 = 7:9)}

\item{.data}{}

\item{.id}{.id = "time",}

\item{names}{Alternative zu den Namen}

\item{.first.data}{Optional wenn nicht by verwendet wird}

\item{values_fill, names_sep, names_vary}{an tidyr::pivot_wider}

\item{by}{Namen der Spalten die behalten werden}
}
\value{
data.frame

tibble
}
\description{
Erweiterung von tidyr::pivot_longer tidyr::pivot_wider

Long_rbind Die Funktion kombiniert zeilenweise und ist eigentlich eine Long-Funktion.
Namen und Labels verden verworfen und nur die Namen der ersten Liste werden behalten
}
\examples{

df <- data.frame(month=rep(1:3,2),
student=rep(c("Amy", "Bob"), each=3),
A=c(9, 7, 6, 8, 6, 9),
B=c(6, 7, 8, 5, 6, 7))

df2<-df |> Wide(student, A, B)



df[-4] |> tidyr::spread(student, A)
df[-4] |> Wide(student, A)

df2
df2  |> Long( A_Amy, A_Bob ,B_Amy, B_Bob, by=~month)





df |>
  tidyr::gather(variable, value, -(month:student)) |>
  tidyr::unite(temp, student, variable) |>
  tidyr::spread(temp, value)
  
 # Wide 
 
dat <- data.frame(
month = rep(1:3, 2),
student = factor(rep(c("Amy", "Bob"), each = 3)),
A = c(19, 27, 16, 28, 10, 29),
B = c(6, 7, 8, 5, 6, 7)
)
#dat |> Wide(student,  A, B)
# dat |> Wide( month ~ student, value="A")
dat |> Wide(student)
dat |> Wide(student,  B)

dat |> Wide(A ~ student)

dat |> Wide(A + B ~ student)
dat |> Wide(A + B ~ student + month)


dat |> Wide(student,  A, B)


dat |> Wide(A ~ student)
dat |> Wide(A ~ student + month)


dat |> Wide(month ~ student, A)
dat |> Wide(month ~ student, A, B)


DF <- data.frame(
id = 1:16,
group = gl(2, 8, labels = c("Control", "Treat")),
age = rnorm(16),
a1 = rnorm(16),
b1 = rnorm(16),
c1 = rnorm(16),
a2 = rnorm(16),
b2 = rnorm(16),
c2 = rnorm(16),
a3 = rnorm(16),
b3 = rnorm(16),
c3 = rnorm(16)
)

Long(
  DF,
  .list = list(
    t0 = c("a1", "b1", "c1"),
    t1 = c("a2", "b2", "c2"),
    t2 = c("a3", "b3", "c3")
  ),
  by =  ~ id + group + age,
  names = c("a", "b", "c")
)
}
