% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot-long.R, R/pivot-wide.R
\name{Long}
\alias{Long}
\alias{Long.formula}
\alias{Long.data.frame}
\alias{Wide}
\alias{Wide.formula}
\alias{Wide.data.frame}
\title{Long und Wide}
\usage{
Long(x, ...)

\method{Long}{formula}(x, data, key = "variable", value = "value", use.label = TRUE, ...)

\method{Long}{data.frame}(
  data,
  ...,
  by = NULL,
  key = "variable",
  value = "value",
  id.vars = all.vars(by),
  use.label = TRUE
)

Wide(...)

\method{Wide}{formula}(x, data, value)

\method{Wide}{data.frame}(data, key, value)
}
\arguments{
\item{x}{data.frame oder formula}

\item{...}{weitere Argument}

\item{data}{data}

\item{key, value}{Namen fuer die Ausgabe}

\item{use.label}{attribut label verwenden}

\item{by, id.vars}{Items}
}
\value{
data.frame
}
\description{
Erweiterung von tidyr::pivot_longer tidyr::pivot_wider

Wide  entspricht tidyr::spread()

Quelle: https://community.rstudio.com/t/spread-with-multiple-value-columns/5378
}
\examples{

df <- data.frame(month=rep(1:3,2),
student=rep(c("Amy", "Bob"), each=3),
A=c(9, 7, 6, 8, 6, 9),
B=c(6, 7, 8, 5, 6, 7))

df2<-df \%>\% Wide(student, c(A, B))



df[-4] \%>\% tidyr::spread(student, A)
df[-4] \%>\% Wide(student, A)


df2  \%>\% Long( Amy_A, Amy_B, Bob_A, Bob_B, by=~month)




df \%>\%
  tidyr::gather(variable, value, -(month:student)) \%>\%
  tidyr::unite(temp, student, variable) \%>\%
  tidyr::spread(temp, value)

  Long( .~ month, df2)
  

#  suppressPackageStartupMessages(library(tidyverse))

dat <- data.frame(
  month = rep(1:3, 2),
  student = factor(rep(c("Amy", "Bob"), each = 3)),
  A = c(9, 7, 6, 8, 6, 9),
  B = c(6, 7, 8, 5, 6, 7)
)

dat \%>\% Wide(student,  c(A, B))
dat \%>\% Wide(student,  c("A", "B"))
dat[-3] \%>\% Wide(student,  B)
dat  \%>\% Wide(student ~ month)
#dat[-3] \%>\% reshape2::dcast(month ~ student)
dat  \%>\% Wide(month ~ student, A)
dat  \%>\% Wide(student ~ month, A)

}
