% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot-long.R, R/pivot-wide.R
\name{Long}
\alias{Long}
\alias{Long.formula}
\alias{Long.data.frame}
\alias{Wide}
\title{Long und Wide}
\usage{
Long(x, ...)

\method{Long}{formula}(x, data, key = "variable", value = "value", use.label = TRUE, ...)

\method{Long}{data.frame}(
  x,
  ...,
  by = NULL,
  key = "variable",
  value = "value",
  id.vars = all.vars(by),
  use.label = TRUE
)

Wide(
  data,
  key,
  ...,
  values_fill = NULL,
  names_sep = "_",
  names_vary = "fastest"
)
}
\arguments{
\item{x}{data.frame oder formula}

\item{...}{key +  values_fro or erster Teil der Formula}

\item{data}{dataframe}

\item{key, value}{Namen fuer die Ausgabe}

\item{use.label}{attribut label verwenden}

\item{by, id.vars}{Items}

\item{values_fill, names_sep, names_vary}{an tidyr::pivot_wider}
}
\value{
data.frame

tibble
}
\description{
Erweiterung von tidyr::pivot_longer tidyr::pivot_wider
}
\examples{

df <- data.frame(month=rep(1:3,2),
student=rep(c("Amy", "Bob"), each=3),
A=c(9, 7, 6, 8, 6, 9),
B=c(6, 7, 8, 5, 6, 7))

df2<-df |> Wide(student, A, B)



df[-4] |> tidyr::spread(student, A)
df[-4] |> Wide(student, A)

df2
df2  |> Long( A_Amy, A_Bob ,B_Amy, B_Bob, by=~month)





df |>
  tidyr::gather(variable, value, -(month:student)) |>
  tidyr::unite(temp, student, variable) |>
  tidyr::spread(temp, value)
  
 # Wide 
 
dat <- data.frame(
month = rep(1:3, 2),
student = factor(rep(c("Amy", "Bob"), each = 3)),
A = c(19, 27, 16, 28, 10, 29),
B = c(6, 7, 8, 5, 6, 7)
)
#dat |> Wide(student,  A, B)
# dat |> Wide( month ~ student, value="A")
dat |> Wide(student)
dat |> Wide(student,  B)

dat |> Wide(A ~ student)

dat |> Wide(A + B ~ student)
dat |> Wide(A + B ~ student + month)


dat |> Wide(student,  A, B)


dat |> Wide(A ~ student)
dat |> Wide(A ~ student + month)


dat |> Wide(month ~ student, A)
dat |> Wide(month ~ student, A, B)

}
