% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string-wrap.R
\name{wrap_string}
\alias{wrap_string}
\alias{wrap_string.character}
\alias{wrap_string.factor}
\alias{wrap_string.data.frame}
\alias{wrap_sentence}
\title{wrap_string (umbrechen)}
\usage{
wrap_string(x, ...)

\method{wrap_string}{character}(
  x,
  width = 25,
  sep = "\\n",
  pattern = NULL,
  replacement = NULL,
  max.lines = NULL,
  max.lines.char = " ..."
)

\method{wrap_string}{factor}(
  x,
  width = 20,
  sep = "\\n",
  pattern = "_",
  replacement = " ",
  max.lines = NULL,
  max.lines.char = "...",
  lvl = levels(x)
)

\method{wrap_string}{data.frame}(
  x,
  width = 20,
  sep = "\\n",
  pattern = "_",
  replacement = " ",
  max.lines = NULL,
  max.lines.char = "..."
)

wrap_sentence(x, ...)
}
\arguments{
\item{x}{data.frame oder String}

\item{width}{default width= 25}

\item{sep}{default new line}

\item{pattern, replacement}{zB Unterstriche _ ersaetzen}

\item{max.lines, max.lines.char}{Anschneiden des Strings}
}
\value{
character

factor

data.frame

character (Labels)
}
\description{
Kopie von  str_wrap  wobei die Labels mitupData2 ergaenzt werden wenn ein 
Data.Frame-Objekt uebergeben wird.

wrap_sentence: Kopie von  str_wrap ruckgabe der Labels
}
\details{
wrap_label kuerzt die label fur Grafiken.
}
\examples{

wrap_string(
  "R is free   software and comes with ABSOLUTELY NO WARRANTY.
  You are welcome to redistribute it under certain conditions.
  ", 5
)




df <- data.frame(
  BMI=c(1,2,3,1,2,3),
  WHtR= gl(2,3, label =c("R is free   software  and comes with ABSOLUTELY NO WARRANTY", 
                         "You are welcome to redistribute it under certain conditions.")),
  WHtR_1=c(9,7,6,8,6,9),
  bildprof=c(6,7,8,5,6,7)
)

DF<-
  Label(df, BMI = "Body-Mass-Index   Masszahl für die Bewertung des Koerpergewichts eines
   Menschen in Relation zu seiner Koerpergroesse.",
        WHtR =  "Waist-Height-Ratio",
        WHtR_1 ="Waist-Height-Ratio"
  )

  # DF$BMI<- units::set_units(DF$BMI, kg/m2)

DF<- wrap_label(DF,   width = 20, max.lines = 1)
get_label(DF)

wrap_factor(DF, max.lines = 1, max.lines.char = "")


}
