% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string-wrap.R
\name{wrap_string}
\alias{wrap_string}
\alias{wrap_string.character}
\alias{wrap_string.factor}
\alias{wrap_string.data.frame}
\alias{wrap_factor}
\alias{wrap_data_label}
\alias{wrap_string_at}
\alias{wrap_string_at.character}
\alias{wrap_string_at.factor}
\alias{split_string}
\alias{split_string.character}
\alias{split_string.factor}
\title{wrap_string (umbrechen)}
\usage{
wrap_string(x, ...)

\method{wrap_string}{character}(
  x,
  width = 25,
  sep = "\\n",
  pattern = NULL,
  replacement = NULL,
  max.lines = NULL,
  max.lines.char = " ...",
  ...
)

\method{wrap_string}{factor}(
  x,
  width = 20,
  sep = "\\n",
  pattern = "_",
  replacement = " ",
  max.lines = NULL,
  max.lines.char = "...",
  lvl = levels(x),
  ...
)

\method{wrap_string}{data.frame}(x, ...)

wrap_factor(
  x,
  width = 20,
  sep = "\\n",
  pattern = "_",
  replacement = " ",
  max.lines = NULL,
  max.lines.char = "...",
  lvl = levels(x),
  ...
)

wrap_data_label(
  x,
  width = 20,
  sep = "\\n",
  pattern = "_",
  replacement = " ",
  max.lines = NULL,
  max.lines.char = "...",
  ...
)

wrap_string_at(x, ...)

\method{wrap_string_at}{character}(x, pattern, replacement = paste0("\\n", pattern), ...)

\method{wrap_string_at}{factor}(
  x,
  pattern,
  replacement = paste0("\\n", pattern),
  lvl = levels(x),
  ...
)

split_string(x, ...)

\method{split_string}{character}(x, pattern, pos = 1, ...)

\method{split_string}{factor}(x, pattern, pos = 1, lvl = levels(x), ...)
}
\arguments{
\item{x}{data.frame oder String}

\item{...}{alles weiter}

\item{width}{default width= 25}

\item{sep}{default new line}

\item{pattern, replacement}{zB Unterstriche _ ersaetzen}

\item{max.lines, max.lines.char}{Anschneiden des Strings}

\item{lvl}{character levels}

\item{pos}{integer Position}
}
\value{
wrap_string: string

character

factor

character (Labels)

wrap_factor(): factor

wrap_data: data.frame

character

factor

character

factor
}
\description{
wrap_string(), wrap_factor(), wrap_data_label(),   wrap_string_at(), split_string()

wrap_sentence: Kopie von  str_wrap ruckgabe der Labels
}
\details{
Kopie von  str_wrap  wobei die Labels mitupData2 ergaenzt werden wenn ein
Data.Frame-Objekt uebergeben wird.
}
\examples{

strg<- c("R is free   software and comes with ABSOLUTELY NO WARRANTY.",
         "You are welcome to redistribute it under certain conditions.")

wrap_string(strg, 5)
wrap_string(factor(strg))
wrap_factor(factor(strg))

#wrap_data_label(data)
wrap_string_at(strg, "and")
split_string(strg, "and")

}
