% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string-wrap.R
\name{wrap_string}
\alias{wrap_string}
\alias{wrap_string.character}
\alias{wrap_string.factor}
\alias{wrap_string.data.frame}
\alias{wrap_sentence}
\alias{wrap_string_at}
\alias{wrap_string_at.character}
\alias{wrap_string_at.factor}
\alias{split_string}
\alias{split_string.character}
\alias{split_string.factor}
\title{wrap_string (umbrechen)}
\usage{
wrap_string(x, ...)

\method{wrap_string}{character}(
  x,
  width = 25,
  sep = "\\n",
  pattern = NULL,
  replacement = NULL,
  max.lines = NULL,
  max.lines.char = " ..."
)

\method{wrap_string}{factor}(
  x,
  width = 20,
  sep = "\\n",
  pattern = "_",
  replacement = " ",
  max.lines = NULL,
  max.lines.char = "...",
  lvl = levels(x)
)

\method{wrap_string}{data.frame}(
  x,
  width = 20,
  sep = "\\n",
  pattern = "_",
  replacement = " ",
  max.lines = NULL,
  max.lines.char = "..."
)

wrap_sentence(x, ...)

wrap_string_at(x, ...)

\method{wrap_string_at}{character}(x, pattern, replacement = paste0("\\n", pattern))

\method{wrap_string_at}{factor}(
  x,
  pattern,
  replacement = paste0("\\n", pattern),
  lvl = levels(x)
)

split_string(x, ...)

\method{split_string}{character}(x, pattern, pos = 1)

\method{split_string}{factor}(x, pattern, pos = 1, lvl = levels(x))
}
\arguments{
\item{x}{data.frame oder String}

\item{...}{alles weiter}

\item{width}{default width= 25}

\item{sep}{default new line}

\item{pattern, replacement}{zB Unterstriche _ ersaetzen}

\item{max.lines, max.lines.char}{Anschneiden des Strings}

\item{lvl}{character levels}

\item{pos}{integer Position}
}
\value{
character

factor

data.frame

character (Labels)

character

factor

character

factor
}
\description{
Kopie von  str_wrap  wobei die Labels mitupData2 ergaenzt werden wenn ein 
Data.Frame-Objekt uebergeben wird.

wrap_sentence: Kopie von  str_wrap ruckgabe der Labels
}
\examples{

wrap_string(
  "R is free   software and comes with ABSOLUTELY NO WARRANTY.
  You are welcome to redistribute it under certain conditions.
  ", 5
)



# example code
  
1+1
# x <- c("Potassium (mg/l)", "Calcium (mg/l)","Adjusted CoQ10 (µmol/mmol Chol)")
# wrap_string_at(x, "\\(")
# wrap_string_at(factor(x), "\\(")

}
